+++ editor/editor-core/src/__tests__/integration/status/selection.ts
@@ -7,6 +7,7 @@ import { getBoundingRect } from '../../__helpers/page-objects/_editor';
 import { STATUS_SELECTORS } from '../../__helpers/page-objects/_status';
 import { fullpage, expectToMatchSelection } from '../_helpers';
 import statusAdf from './__fixtures__/status-single.adf.json';
+import statusWithTextAdf from './__fixtures__/status-with-text.adf.json';
 
 BrowserTestCase(
   'selection.ts: Clicking after a status produces a text selection to its right',
@@ -31,6 +32,38 @@ BrowserTestCase(
     await expectToMatchSelection(page, {
       anchor: positionAfterStatus,
       head: positionAfterStatus,
+      type: 'text',
+    });
+  },
+);
+
+BrowserTestCase(
+  'selection.ts: selecting line with a status and text via keys',
+  {},
+  async (client: any, testName: string) => {
+    const page = await goToEditorTestingWDExample(client);
+    await mountEditor(page, {
+      appearance: fullpage.appearance,
+      allowStatus: true,
+      defaultValue: statusWithTextAdf,
+    });
+    await page.waitForSelector(STATUS_SELECTORS.STATUS_NODE);
+
+    // click in the text after the status and move cursor to the end of the line
+    const textAfterStatusSelector = `//p[contains(text(), "text after status")]`;
+    await page.click(textAfterStatusSelector);
+    await page.keys(['End'], true);
+    await expectToMatchSelection(page, {
+      anchor: 19,
+      head: 19,
+      type: 'text',
+    });
+
+    // select the whole line with keys
+    await page.keys(['Shift', 'Home'], true);
+    await expectToMatchSelection(page, {
+      anchor: 19,
+      head: 1,
       type: 'text',
     });
   },