+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 29.1.4
+
+### Patch Changes
+
+- Updated dependencies
+
+## 29.1.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 29.1.2
 
 ### Patch Changes